+++ python/lsst/ts/mtdome/mtdome_csc.py
@@ -19,7 +19,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["MTDomeCsc"]
+__all__ = ["MTDomeCsc", "DOME_AZIMUTH_OFFSET", "run_mtdome"]
 
 import asyncio
 import math
@@ -31,6 +31,7 @@ from . import __version__, encoding_tools
 from .llc_configuration_limits import AmcsLimits, LwscsLimits
 from .enums import LlcMotionState, LlcName, ResponseCode, LlcNameDict
 from lsst.ts import salobj
+from lsst.ts import utils
 from .mock_controller import MockMTDomeController
 from lsst.ts.idl.enums.MTDome import (
     EnabledState,
@@ -40,25 +41,72 @@ from lsst.ts.idl.enums.MTDome import (
 )
 
 _LOCAL_HOST = "127.0.0.1"
-_TIMEOUT = 20  # timeout in s to be used by this module
+_TIMEOUT = 20  # timeout [sec] to be used by this module
 _KEYS_TO_REMOVE = {
     "status",
-    "operationalMode",  # Remove when next XML release is done
-    "appliedConfiguration",  # Remove when next XML release is done
-}
-_KEYS_IN_RADIANS = {
-    "positionActual",
-    "positionCommanded",
-    "velocityActual",
-    "velocityCommanded",
+    "operationalMode",  # Remove because gets emitted as an event
+    "appliedConfiguration",  # Remove because gets emitted as an event
 }
 
+# The values of these keys need to be compensated for the dome azimuth offset
+# in the AMCS status. Note that these keys are shared with LWSCS so they can be
+# added to _KEYS_IN_RADIANS to avoid duplication but this also means that an
+# additional check for the AMCS lower level component needs to be done when
+# applying the offset correction. That is a trade off I can live with.
+_AMCS_KEYS_OFFSET = {
+    "positionActual",
+    "positionCommanded",
+}
+# The values of these keys need to be converted from radians to degrees when
+# the status is recevied and telemetry with these values is sent.
+_KEYS_IN_RADIANS = {
+    "velocityActual",
+    "velocityCommanded",
+}.union(_AMCS_KEYS_OFFSET)
+
+# The offset of the dome rotation zero point with respect to azimuth 0º (true
+# north) is 32º west and the aperture lies at the opposite side of the dome.
+# Therefore the dome azimuth offset is 180º - 32º = 148º and this needs to be
+# subtracted when commanding the azimuth position, or added when sending the
+# azimuth telemetry.
+DOME_AZIMUTH_OFFSET = 148.0
+
+# Polling periods [sec] for the lower level components.
 _AMCS_STATUS_PERIOD = 0.2
 _APsCS_STATUS_PERIOD = 2.0
 _LCS_STATUS_PERIOD = 2.0
 _LWSCS_STATUS_PERIOD = 2.0
 _MONCS_STATUS_PERIOD = 2.0
 _THCS_STATUS_PERIOD = 2.0
+
+# These next commands are temporarily disabled because they will be issued
+# during the upcoming TMA pointing test and the EIE LabVIEW code doesn't handle
+# them yet, which will result in an error. As soon as the TMA pointing test is
+# done, they will be reenabled. The name reflects the fact that there probably
+# will be more situations during commissioning in which commands need to be
+# disabled.
+COMMANDS_DISABLED_FOR_COMMISSIONING = {
+    "closeLouvers",
+    "closeShutter",
+    "crawlEl",
+    "fans",
+    "goStationaryEl",
+    "goStationaryLouvers",
+    "goStationaryShutter",
+    "inflate",
+    "moveEl",
+    "openShutter",
+    "setLouvers",
+    "setTemperature",
+    "stopEl",
+    "stopLouvers",
+    "stopShutter",
+}
+REPLY_DATA_FOR_DISABLED_COMMANDS = {"response": 0, "timeout": 0}
+
+
+def run_mtdome() -> None:
+    asyncio.run(MTDomeCsc.amain(index=None))
 
 
 class MTDomeCsc(salobj.ConfigurableCsc):
@@ -151,26 +199,29 @@ class MTDomeCsc(salobj.ConfigurableCsc):
                 OperationalMode.NORMAL.name: "setNormalAz",
                 OperationalMode.DEGRADED.name: "setDegradedAz",
             },
-            SubSystemId.LWSCS: {
-                OperationalMode.NORMAL.name: "setNormalEl",
-                OperationalMode.DEGRADED.name: "setDegradedEl",
-            },
-            SubSystemId.APSCS: {
-                OperationalMode.NORMAL.name: "setNormalShutter",
-                OperationalMode.DEGRADED.name: "setDegradedShutter",
-            },
-            SubSystemId.LCS: {
-                OperationalMode.NORMAL.name: "setNormalLouvers",
-                OperationalMode.DEGRADED.name: "setDegradedLouvers",
-            },
-            SubSystemId.MONCS: {
-                OperationalMode.NORMAL.name: "setNormalMonitoring",
-                OperationalMode.DEGRADED.name: "setDegradedMonitoring",
-            },
-            SubSystemId.THCS: {
-                OperationalMode.NORMAL.name: "setNormalThermal",
-                OperationalMode.DEGRADED.name: "setDegradedThermal",
-            },
+            # The next lines have been commented out because the systems
+            # concerned are not available at the summit yet. As soon as they
+            # are made available, the corresponding lines will be uncommented.
+            # SubSystemId.LWSCS: {
+            #     OperationalMode.NORMAL.name: "setNormalEl",
+            #     OperationalMode.DEGRADED.name: "setDegradedEl",
+            # },
+            # SubSystemId.APSCS: {
+            #     OperationalMode.NORMAL.name: "setNormalShutter",
+            #     OperationalMode.DEGRADED.name: "setDegradedShutter",
+            # },
+            # SubSystemId.LCS: {
+            #     OperationalMode.NORMAL.name: "setNormalLouvers",
+            #     OperationalMode.DEGRADED.name: "setDegradedLouvers",
+            # },
+            # SubSystemId.MONCS: {
+            #     OperationalMode.NORMAL.name: "setNormalMonitoring",
+            #     OperationalMode.DEGRADED.name: "setDegradedMonitoring",
+            # },
+            # SubSystemId.THCS: {
+            #     OperationalMode.NORMAL.name: "setNormalThermal",
+            #     OperationalMode.DEGRADED.name: "setDegradedThermal",
+            # },
         }
 
         self.log.info("DomeCsc constructed")
@@ -196,6 +247,7 @@ class MTDomeCsc(salobj.ConfigurableCsc):
             host = self.config.host
             port = self.config.port
         try:
+            self.log.info(f"Connecting to host={host} and port={port}")
             connect_coro = asyncio.open_connection(host=host, port=port)
             self.reader, self.writer = await asyncio.wait_for(
                 connect_coro, timeout=self.config.connection_timeout
@@ -328,35 +380,33 @@ class MTDomeCsc(salobj.ConfigurableCsc):
         command_dict = dict(command=command, parameters=params)
         st = encoding_tools.encode(**command_dict)
         async with self.communication_lock:
-            self.log.debug(f"Sending command {st}")
             try:
                 assert self.writer is not None
             except AssertionError as e:
                 await self.fault(code=3, report=f"Error writing command {st}: {e}")
                 raise
-            self.writer.write(st.encode() + b"\r\n")
-            await self.writer.drain()
-            try:
-                assert self.reader is not None
-                read_bytes = await asyncio.wait_for(
-                    self.reader.readuntil(b"\r\n"), timeout=_TIMEOUT
+
+            if command not in COMMANDS_DISABLED_FOR_COMMISSIONING:
+                self.writer.write(st.encode() + b"\r\n")
+                await self.writer.drain()
+                try:
+                    assert self.reader is not None
+                    read_bytes = await asyncio.wait_for(
+                        self.reader.readuntil(b"\r\n"), timeout=_TIMEOUT
+                    )
+                except (asyncio.exceptions.TimeoutError, AssertionError) as e:
+                    await self.fault(
+                        code=3, report=f"Error reading reply to command {st}: {e}"
+                    )
+                    raise
+                data: typing.Dict[str, typing.Any] = encoding_tools.decode(
+                    read_bytes.decode()
                 )
-            except (asyncio.exceptions.TimeoutError, AssertionError) as e:
-                await self.fault(
-                    code=3, report=f"Error reading reply to command {st}: {e}"
-                )
-                raise
-            data: typing.Dict[str, typing.Any] = encoding_tools.decode(
-                read_bytes.decode()
-            )
+            else:
+                data = REPLY_DATA_FOR_DISABLED_COMMANDS
             response = data["response"]
 
-            if response == ResponseCode.OK:
-                if "status" not in command:
-                    self.log.info(f"Received reply {data}")
-                else:
-                    self.log.debug(f"Received reply {data}")
-            else:
+            if response != ResponseCode.OK:
                 self.log.error(f"Received ERROR {data}.")
                 if response == ResponseCode.COMMAND_REJECTED:
                     raise ValueError(f"The command {command} was rejected.")
@@ -377,9 +427,13 @@ class MTDomeCsc(salobj.ConfigurableCsc):
         self.log.debug(
             f"Moving Dome to azimuth {data.position} and then start crawling at azRate {data.velocity}"
         )
+        # Compensate for the dome azimuth offset.
+        position = utils.angle_wrap_nonnegative(
+            data.position - DOME_AZIMUTH_OFFSET
+        ).degree
         await self.write_then_read_reply(
             command="moveAz",
-            position=math.radians(data.position),
+            position=math.radians(position),
             velocity=math.radians(data.velocity),
         )
         await self.evt_azTarget.set_write(
@@ -777,9 +831,9 @@ class MTDomeCsc(salobj.ConfigurableCsc):
             command=command
         )
 
-        # DM-30807: Send OperationalMode event at start up.
-        current_operational_mode = status[llc_name]["status"]["operationalMode"]
         if llc_name not in self.lower_level_status:
+            # DM-30807: Send OperationalMode event at start up.
+            current_operational_mode = status[llc_name]["status"]["operationalMode"]
             operatinal_mode = OperationalMode[current_operational_mode]
             sub_system_id = [
                 sid for sid, name in LlcNameDict.items() if name == llc_name
@@ -789,6 +843,16 @@ class MTDomeCsc(salobj.ConfigurableCsc):
                 subSystemId=sub_system_id,
             )
 
+            # DM-34664: Send appliedConfiguration event as well, if present.
+            if "appliedConfiguration" in status[llc_name]:
+                applied_configuration = status[llc_name]["appliedConfiguration"]
+                jmax = applied_configuration["jmax"]
+                amax = applied_configuration["amax"]
+                vmax = applied_configuration["vmax"]
+                await self.evt_azConfigurationApplied.set_write(
+                    jmax=jmax, amax=amax, vmax=vmax
+                )
+
         # Store the status for reference.
         self.lower_level_status[llc_name] = status[llc_name]
 
@@ -797,6 +861,16 @@ class MTDomeCsc(salobj.ConfigurableCsc):
         for key in telemetry_in_radians.keys():
             if key in _KEYS_IN_RADIANS and llc_name in [LlcName.AMCS, LlcName.LWSCS]:
                 telemetry_in_degrees[key] = math.degrees(telemetry_in_radians[key])
+                # Compensate for the dome azimuth offset. This is done here and
+                # not one level higher since angle_wrap_nonnegative only
+                # accepts Angle or a float in degrees and this way the
+                # conversion from radians to degrees only is done in one line
+                # of code.
+                if key in _AMCS_KEYS_OFFSET and llc_name == LlcName.AMCS:
+                    offset_value = utils.angle_wrap_nonnegative(
+                        telemetry_in_degrees[key] + DOME_AZIMUTH_OFFSET
+                    ).degree
+                    telemetry_in_degrees[key] = offset_value
             elif key == "timestampUTC":
                 # DM-26653: The name of this parameter is still under
                 # discussion.
@@ -843,7 +917,7 @@ class MTDomeCsc(salobj.ConfigurableCsc):
                 # In case of some unit tests, this event is expected to be
                 # emitted twice with the same data.
                 await self.evt_azMotion.set_write(
-                    force_output=True, state=motion_state, inPosition=in_position
+                    state=motion_state, inPosition=in_position
                 )
         elif llc_name == LlcName.LWSCS:
             llc_status = status[llc_name]["status"]